+++ mycodo/inputs/scd4x_circuitpython.py
@@ -52,7 +52,7 @@ INPUT_INFORMATION = {
     'dependencies_module': [
         ('pip-pypi', 'usb.core', 'pyusb==1.1.1'),
         ('pip-pypi', 'adafruit_extended_bus', 'Adafruit-extended-bus==1.0.2'),
-        ('pip-pypi', 'adafruit_scd4x', 'adafruit-circuitpython-scd4x==1.2.2')
+        ('pip-pypi', 'adafruit_scd4x', 'adafruit-circuitpython-scd4x==1.3.8')
     ],
 
     'interfaces': ['I2C'],